move tooltip component to common folder
