+++ bigbluebutton-html5/imports/ui/components/presentation/presentation-toolbar/component.jsx
@@ -6,7 +6,7 @@ import injectWbResizeEvent from '/imports/ui/components/presentation/resize-wrap
 import { HUNDRED_PERCENT, MAX_PERCENT, STEP } from '/imports/utils/slideCalcUtils';
 import Styled from './styles';
 import ZoomTool from './zoom-tool/component';
-import TooltipContainer from '/imports/ui/components/tooltip/container';
+import TooltipContainer from '/imports/ui/components/common/tooltip/container';
 import KEY_CODES from '/imports/utils/keyCodes';
 
 const intlMessages = defineMessages({
